+++ src/main/java/com/ciandt/model/package-info.java
@@ -1,0 +1,10 @@
+/**
+  * Model classes.
+  * <p>
+  * These classes implement the models for data habdling.
+  * </p>
+  *
+  * @author Helio Diniz 
+  * @version 1.0
+  */
+package com.ciandt.model;