+++ src/main/java/com/ciandt/repository/package-info.java
@@ -1,0 +1,10 @@
+/**
+  * Repository classes.
+  * <p>
+  * These classes implement the application's data repositories.
+  * </p>
+  *
+  * @author Helio Diniz 
+  * @version 1.0
+  */
+package com.ciandt.repository;