+++ src/main/java/com/ciandt/service/package-info.java
@@ -1,0 +1,10 @@
+/**
+  * Service classes.
+  * <p>
+  * These classes implements the services provided by the application.
+  * </p>
+  *
+  * @author Helio Diniz 
+  * @version 1.0
+  */
+package com.ciandt.service;