+++ src/main/java/com/ciandt/utils/package-info.java
@@ -1,0 +1,10 @@
+/**
+  * Utility classes.
+  * <p>
+  * These classes implements tools used by the application.
+  * </p>
+  *
+  * @author Helio Diniz 
+  * @version 1.0
+  */
+package com.ciandt.utils;