+++ src/test/java/com/ciandt/json/test/package-info.java
@@ -1,0 +1,10 @@
+/**
+  * JSON test classes.
+  * <p>
+  * These classes implements the JSON utilities tests.
+  * </p>
+  *
+  * @author Helio Diniz 
+  * @version 1.0
+  */
+package com.ciandt.json.test;